Provide a more complete default configuration file.
Summary: Actually, this is the 2nd half of D24. As mentioned there, this is the solution to getting the {key Super} key to toggle the menu.

Test Plan: Um, press the {key Super} key and make sure the menu opens. ☺ To be clear, this is konwn to work, as this path is hardcoded and it's the reason why the file we removed in D24 doesn't work.

Reviewers: tsimonq2

Reviewed By: tsimonq2

Differential Revision: https://phab.lubuntu.me/D31